Create Perfect Number
A perfect number is a positive integer that is equal to the sum of its proper divisors (excluding the number itself). In other words, for a number n, if the sum of all positive divisors of n (excluding n itself) equals n, then  n is a Perfect number.

Example of Perfect Numbers:
6: The divisors of 6 (excluding 6) are 1, 2, and 3. Their sum is:
1+2+3=6
So, 6 is a perfect number.